fix: extracts port from host
in addition moved integration tests to its own folder so we can add unit tests to the ossm installer without a need of starting kube testenv. We cannot use label in Ginkgo as we do in project controller because we are stuck with v1 of test framework. Updating to v2 causes a lot of depedency issues